requirements.txt
Django==4.2.23
djangorestframework==3.15.2
graphene-django==3.2.3
django-graphql-jwt==0.4.0
graphene-mongo
mongoengine==0.29.1
django-cors-headers==4.4.0
python-decouple==3.8
bcrypt==4.0.1
pymongo==4.10.1
dnspython==2.6.1 
